Hacked i-v3.s to skip loading into ram and reu completely
The plan is to implement a .z3 filetype for the ultimate that can
be run right from the ultimate menu. The ultimate will load the
interpreter and the resident code directly into RAM using DMA and
preload the rest of the story file into the REU, then simply run
everything.

Therefore the relevant parts are simply skipped and the interpreter
starts to run the story right away.